User has been registration fixed.

+++ it4kids/src/main/java/it4kids/domain/AbstractModel.java
@@ -10,8 +10,8 @@ package it4kids.domain;
 public abstract class AbstractModel {
 	private long id;
 
-	public long getId() {
-		return id;
+	public int getId() {
+		return (int) id;
 	}
 
 	public void setId(long id) {
